Since Treemap.tsx contains state, it should be considered a Container; Moving it to container directory.
